Домашнее задание 23го спринта: Финал
Сверстан экран плейлиста
Реализовано вывод информации о плейлисте на этом экране
Изменена иконка приложения

+++ app/src/main/java/com/example/playlistmaker/search/data/SearchRepositoryImpl.kt
@@ -53,7 +53,7 @@ class SearchRepositoryImpl(
                                 trackId,
                                 trackName,
                                 artistName,
-                                trackTimeMillis,
+                                trackTimeMillis.toLong(),
                                 artworkUrl100,
                                 collectionName,
                                 releaseDate,
